Updating PaymentLog model
Instead of order_id, order_number is now displayed in the list of payment logs. This solves the problem of mismatch of order numbers in the list of orders and payment logs if the id and order_number are different in the Order model (for some reason, I had an out of sync id and order_id.).