Event after order create and before cart delete
I've a situation where i'd like to collect an extra field (delivery date), from the cart and assign it to the order upon creation. Since the order creation is within a transaction, and the cart gets deleted within the transaction, I can't find a way to transfer elements from one to the other.

Conceptually, I'm requesting an event to be fired within the transaction that passes both the recently created order and the soon to be deleted cart. This would allow adding any number of extra fields to a cart and manipulating the resulting order how the developer sees fit. 

In any Plugin.php, one would:

```php
Event::listen('mall.order.afterCreate', function($order, $cart){
    $order->extraField = $cart->extraField;
    
    // or
    
    $order->existingAttribute = $cart->newManipulatingMethod();
});
```
You get the idea.