Only show published variants in menus
Currently "[OFFLINE.Mall] All Shop Variants" type menu items show published variants only, which is correct. But they also show published variants of products which have not been published. This is incorrect. This fix aims to correct this by also filtering on the published property of the product related to each variant.